Working on adding resume parsing and correcting parsing for new form

+++ .parser/scrape.rb
@@ -5,37 +5,40 @@
 require 'wuparty'
 require 'fileutils'
 
-def get_api_key
-  key = ""
+ACCOUNT = 'flatironschool'
 
-  File.open("api_key.txt", "r").each_line do |line|
-    key = line
+def get_api_key
+  key = { }
+
+  count = 0
+  File.open("wufoo_info.txt", "r").each_line do |line|
+    line = line.gsub(/\n/, "")
+    key[count] = line
+    count += 1
   end
 
   key
 end
 
-ACCOUNT = 'flatironschool'
-FORM_ID = 'ruby-002-student-profile-survey'
-
 def scraper
 
   names = []
-
-  api_key = get_api_key
+  wufoo_info = get_api_key
+  api_key = wufoo_info[0]
+  form_id = wufoo_info[1]
 
   if api_key == ""
     print "\nNo API key set. Please run `./parse.sh api [key]`.\n"
     return 0
-  elsif api_key.length != 19
-    print "\nAPI key is invalid.\n"
-    return 0
   end
-
+  # And other error-catching for other fields and other possible errors.
+  # (check response from Wufoo before?)
   wufoo = WuParty.new(ACCOUNT, api_key)
-  form = wufoo.form(FORM_ID)
+  form = wufoo.form(form_id)
 
-  FileUtils.mv('bak/sps.txt', 'bak/sps.txt.bak')
+  if File.exist?('bak/sps.txt')
+    FileUtils.mv('bak/sps.txt', 'bak/sps.txt.bak')
+  end
 
   File.open("bak/sps.txt", "w") do |sps|
 
@@ -56,18 +59,6 @@ def scraper
       end
 
       names << name
-
-
-      # for profile_looper in name
-      #   Page_Titles << profile_looper.to_s
-      # end
-      # Page_Titles << "_profile\n"
-      # for resume_looper in name
-      #   Page_Titles << resume_looper.to_s
-      # end
-      # Page_Titles << "_resume\n"
-
-
       txt << "\n"
       sps.write(txt)
